+++ src/cdcprolific.cpp
@@ -26,7 +26,7 @@ uint32_t PL2303::Init(uint32_t parent, uint32_t port, uint32_t lowspeed) {
 
         uint8_t buf[constBufSize];
         USB_DEVICE_DESCRIPTOR * udd = reinterpret_cast<USB_DEVICE_DESCRIPTOR*>(buf);
-        uint8_t rcode;
+        uint32_t rcode;
         UsbDeviceDefinition *p = NULL;
         EpInfo *oldep_ptr = NULL;
         uint8_t num_of_conf; // number of configurations
@@ -232,7 +232,7 @@ Fail:
 
 //uint8_t PL::Poll()
 //{
-//      uint8_t rcode = 0;
+//      uint32_t rcode = 0;
 //
 //      //if (!bPollEnable)
 //      //      return 0;
