+++ src/hidcomposite.cpp
@@ -97,7 +97,7 @@ uint32_t HIDComposite::Init(uint32_t parent, uint32_t port, uint32_t lowspeed) {
 
         uint8_t buf[constBufSize];
         USB_DEVICE_DESCRIPTOR * udd = reinterpret_cast<USB_DEVICE_DESCRIPTOR*>(buf);
-        uint8_t rcode;
+        uint32_t rcode;
         UsbDeviceDefinition *p = NULL;
         EpInfo *oldep_ptr = NULL;
         uint8_t len = 0;
@@ -355,7 +355,7 @@ void HIDComposite::ZeroMemory(uint8_t len, uint8_t *buf) {
 }
 
 uint32_t HIDComposite::Poll() {
-        uint8_t rcode = 0;
+        uint32_t rcode = 0;
 
         if(!bPollEnable)
                 return 0;
@@ -375,7 +375,7 @@ uint32_t HIDComposite::Poll() {
 
                         ZeroMemory(constBuffLen, buf);
 
-                        uint8_t rcode = pUsb->inTransfer(bAddress, epInfo[index].epAddr, &read, buf);
+                        uint32_t rcode = pUsb->inTransfer(bAddress, epInfo[index].epAddr, &read, buf);
 
                         if(rcode) {
                                 if(rcode != USB_ERRORFLOW)
